+++ homeworks/week13/hw3/src/js/variables.js
@@ -1,0 +1,5 @@
+export const ROOTURL = 'https://api.twitch.tv/kraken/'
+export const CLIENTID = 'r0me6woz37936skc3tsuviuipkp9mb'
+export const ACCEPT = 'application/vnd.twitchtv.v5+json'
+export const GAMELIMIT = 5
+export const STREAMLIMIT = 20
