+++ homeworks/week13/hw3/src/style.css
@@ -89,7 +89,7 @@ html {
 }
 
 .background {
-  background-image: url(./src/background.jpg);
+  background-image: url(./background.jpg);
   width: 100%;
   height: 100%;
   background-position: 50% 50%;
